Add the choice for detection rules

+++ 01_deploy.sh
@@ -9,6 +9,7 @@ echo "###### CONFIGURING ACCOUNT ELASTIC #######"
 echo "###### AND KIBANA API KEY          ######"
 echo "##########################################"
 echo  
+echo
 password=$(< /dev/urandom tr -dc A-Za-z0-9 | head -c14)
 kibana_password=$(< /dev/urandom tr -dc A-Za-z0-9 | head -c14)
 kibana_api_key=$(< /dev/urandom tr -dc A-Za-z0-9 | head -c32)
@@ -16,6 +17,7 @@ cortex_api=$(< /dev/urandom tr -dc A-Za-z0-9 | head -c32)
 echo "The master password Elastic set in .env:" $password
 echo "The master password Kibana set in .env:" $kibana_password
 echo "The Kibana api key is : " $kibana_api_key
+echo
 echo
 sed -i "s|kibana_api_key|$kibana_api_key|g" kibana/kibana.yml
 sed -i "s|changeme|$password|g" .env cortex/application.conf thehive/application.conf elastalert/elastalert.yaml filebeat/filebeat.yml metricbeat/metricbeat.yml heartbeat/heartbeat.yml metricbeat/modules.d/elasticsearch-xpack.yml metricbeat/modules.d/kibana-xpack.yml kibana/kibana.yml auditbeat/auditbeat.yml logstash/config/logstash.yml logstash/pipeline/beats/300_output_beats.conf logstash/pipeline/stoq/300_output_stoq.conf sigma/dockerfile arkime/scripts/capture.sh arkime/scripts/config.sh arkime/scripts/import.sh arkime/scripts/init-db.sh arkime/scripts/viewer.sh arkime/config.ini cortex/Elasticsearch_IP.json cortex/Elasticsearch_Hash.json
@@ -78,10 +80,10 @@ echo "### CONFIGURING CLUSTER ELASTICSEARCH  ###"
 echo "##########################################"
 echo
 echo
-read -p "Enter the RAM of master node elasticsearch [2]: " master_node
+read -p "Enter the RAM in Go of master node elasticsearch [2]: " master_node
 master_node=${master_node:-2}
 sed -i "s|RAM_MASTER|$master_node|g" docker-compose.yml
-read -p "Enter the RAM of data,ingest node elasticsearch [4]: " data_node
+read -p "Enter the RAM in Go of data,ingest node elasticsearch [4]: " data_node
 data_node=${data_node:-4}
 sed -i "s|RAM_DATA|$data_node|g" docker-compose.yml
 echo
@@ -111,6 +113,23 @@ systemctl enable S1EM-promiscuous
 # start service
 systemctl start S1EM-promiscuous
 ###########
+echo
+echo
+echo "##########################################"
+echo "########## CONFIGURE DETECTION ###########"
+echo "##########################################"
+echo
+echo
+while true; do
+    read -r -p "Do you want use detection with rules of Sigma or Elastic (Elastic recommanded) [E/S]?" rules
+    case $rules in
+        [Ee]) detection=ELASTIC; break;;
+        [Ss]) detection=SIGMA; break;;
+        * ) echo "Please answer (E/e) or (S/s).";;
+    esac
+done
+echo
+echo "$detection"
 echo
 echo
 echo "##########################################"
@@ -419,15 +438,23 @@ docker restart cortex
 echo
 echo
 echo "##########################################"
-echo "######### INSTALL SIGMA RULES ############"
+echo "####### INSTALL DETECTION RULES ##########"
 echo "##########################################"
 echo
 echo
 curl -sk -XPOST -u elastic:$password "https://127.0.0.1/kibana/s/default/api/detection_engine/index" -H "kbn-xsrf: true"
-docker image rm -f sigma:1.0
-docker container prune -f
-docker-compose -f sigma.yml build
-docker-compose -f sigma.yml up -d
+echo
+echo
+if 	 [ "$detection" == ELASTIC ];
+then
+        curl -sk -XPUT -u elastic:$password "https://127.0.0.1/kibana/s/default/api/detection_engine/rules/prepackaged" -H "kbn-xsrf: true"
+elif [ "$detection" == SIGMA ];
+then
+        docker image rm -f sigma:1.0
+        docker container prune -f
+        docker-compose -f sigma.yml build
+        docker-compose -f sigma.yml up -d
+fi
 echo
 echo
 echo "#########################################"
